# apps/api/app/Models/Prodi.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Prodi extends Model
{
    use HasFactory;

    protected $fillable = [
        'code',
        'name',
        'degree',
        'is_active',
    ];

    protected $casts = [
        'is_active' => 'boolean',
    ];
}